requirements.txt
Django==3.1.6
djangorestframework==3.12.2
django-cors-headers==3.7.0
django-oauth-toolkit==1.4.0
djangorestframework-oauth==1.1.0
requests==2.24.0
pyyaml==5.4.1
Pillow==8.1.0
https://github.com/crs4/ome_seadragon_cache/archive/v0.2.0.zip